requirements/dev.txt
# -r test.txt

# Debug tools
bpython==0.13
django-debug-panel==0.8.1
django-debug-toolbar==1.3.2
django-pdb==0.3.2
sqlparse==0.1.10

# runserver_plus
Werkzeug==0.9.6

git+https://github.com/edx/xblock-sdk.git@0.1.1#egg=xblock-sdk==0.1.1